minor fixes

+++ TyranoKurwusBot/Core/Downloaders/VideoMetdata.cs
@@ -6,7 +6,7 @@ public record VideoMetadata(
     string Id,
     string Extractor,
     string Title,
-    int Duration,
+    double Duration,
     (string Id, string Url)[] Thumbnails,
     VideoMetadata?[]? Entries,
     [property: JsonPropertyName("original_url")]
